refactor: Update post views to use post_queryset for consistency

+++ src/blog/views.py
@@ -4,6 +4,8 @@ from django.contrib.auth.decorators import login_required
 from .models import Post
 from .forms import PostForm
 
+post_queryset = Post.objects.all()
+
 
 @login_required
 def post_create(request):
@@ -12,31 +14,44 @@ def post_create(request):
         if form.is_valid():
             form.instance.author = request.user
             form.save()
-            return redirect("blog:detail", pk=form.instance.pk)
+            return redirect("blog:detail", pk=form.instance.id)
     return render(request, "blog/post_create.html", {"form": form})
 
 
 def post_list(request):
-    posts = Post.objects.all()
+    posts = post_queryset.all()
     return render(request, "blog/post_list.html", {"posts": posts})
 
 
 def post_detail(request, pk):
-    post = Post.objects.get(pk=pk)
+    posts = post_queryset.filter(pk=pk)
+    if not posts.exists():
+        return redirect("blog:list")
+
+    post = posts.first()
     return render(request, "blog/post_detail.html", {"post": post})
 
 
+@login_required
 def post_update(request, pk):
-    post = Post.objects.get(pk=pk)
+    posts = post_queryset.filter(pk=pk, author=request.user)
+    if not posts.exists():
+        return redirect("blog:list")
+
+    post = posts.first()
     form = PostForm(request.POST or None, instance=post)
+
     if request.method == "POST":
         if form.is_valid():
             form.save()
-            return redirect("blog:detail", pk=post.pk)
+            return redirect("blog:detail", pk=post.id)
     return render(request, "blog/post_update.html", {"form": form})
 
 
+@login_required
 def post_delete(request, pk):
-    post = Post.objects.get(pk=pk)
-    post.delete()
+    posts = post_queryset.filter(pk=pk, author=request.user)
+    if posts.exists():
+        post = posts.first()
+        post.delete()
     return redirect("blog:list")
